requirements/requirements.txt
torch
torchvision
torchaudio
tqdm
ninja
h5py
kornia
matplotlib
pandas
sklearn
scipy
seaborn
wandb
PyYaml
click
requests
pyspng
imageio-ffmpeg
timm
opencv-python
medmnist
numpy
pandas
plotly
optuna
ipython
pillow
openpyxl
natsort
pydicom
bokeh
umap-learn
torchmetrics
python-gdcm